Move getIdFromRoute from cf.helpers to utils.service
