+++ src/frontend/packages/cloud-foundry/src/features/cloud-foundry/cf.helpers.ts
@@ -3,7 +3,6 @@ import { Store } from '@ngrx/store';
 import { combineLatest, Observable } from 'rxjs';
 import { filter, first, map, publishReplay, refCount, tap } from 'rxjs/operators';
 
-import { CFEntityConfig } from '../../cf-types';
 import { CFAppState } from '../../../../cloud-foundry/src/cf-app-state';
 import { getCFEntityKey } from '../../../../cloud-foundry/src/cf-entity-helpers';
 import { applicationEntityType } from '../../../../cloud-foundry/src/cf-entity-types';
@@ -24,14 +23,14 @@ import { IServiceInstance, IUserProvidedServiceInstance } from '../../../../core
 import { ISpace } from '../../../../core/src/core/cf-api.types';
 import { CurrentUserPermissions } from '../../../../core/src/core/current-user-permissions.config';
 import { CurrentUserPermissionsService } from '../../../../core/src/core/current-user-permissions.service';
-import { pathGet } from '../../../../core/src/core/utils.service';
+import { getIdFromRoute, pathGet } from '../../../../core/src/core/utils.service';
 import {
   extractActualListEntity,
 } from '../../../../core/src/shared/components/list/data-sources-controllers/local-filtering-sorting';
-import { MultiActionListEntity } from '../../../../store/src/monitors/pagination-monitor';
-import { PaginationMonitorFactory } from '../../../../store/src/monitors/pagination-monitor.factory';
 import { SetClientFilter } from '../../../../store/src/actions/pagination.actions';
 import { RouterNav } from '../../../../store/src/actions/router.actions';
+import { MultiActionListEntity } from '../../../../store/src/monitors/pagination-monitor';
+import { PaginationMonitorFactory } from '../../../../store/src/monitors/pagination-monitor.factory';
 import { getPaginationObservables } from '../../../../store/src/reducers/pagination-reducer/pagination-reducer.helper';
 import { endpointEntitiesSelector } from '../../../../store/src/selectors/endpoint.selectors';
 import { selectPaginationState } from '../../../../store/src/selectors/pagination.selectors';
@@ -39,6 +38,7 @@ import { APIResource } from '../../../../store/src/types/api.types';
 import { EndpointModel } from '../../../../store/src/types/endpoint.types';
 import { PaginatedAction, PaginationEntityState } from '../../../../store/src/types/pagination.types';
 import { cfEntityFactory } from '../../cf-entity-factory';
+import { CFEntityConfig } from '../../cf-types';
 import { ActiveRouteCfCell, ActiveRouteCfOrgSpace } from './cf-page.types';
 
 export interface IUserRole<T> {
@@ -215,15 +215,6 @@ export const getRowMetadata = (entity: APIResource | MultiActionListEntity) => {
   }
   return entity.metadata ? entity.metadata.guid : null;
 };
-
-export function getIdFromRoute(activatedRoute: ActivatedRoute, id: string) {
-  if (activatedRoute.snapshot.params[id]) {
-    return activatedRoute.snapshot.params[id];
-  } else if (activatedRoute.parent) {
-    return getIdFromRoute(activatedRoute.parent, id);
-  }
-  return null;
-}
 
 export function getActiveRouteCfOrgSpace(activatedRoute: ActivatedRoute) {
   return ({
