+++ src/frontend/packages/cloud-foundry/src/shared/components/add-service-instance/csi-mode.service.ts
@@ -5,9 +5,9 @@ import { filter, map } from 'rxjs/operators';
 
 import { CFAppState } from '../../../../../cloud-foundry/src/cf-app-state';
 import { serviceBindingEntityType } from '../../../../../cloud-foundry/src/cf-entity-types';
-import { getIdFromRoute } from '../../../../../cloud-foundry/src/features/cloud-foundry/cf.helpers';
 import { SpaceScopedService } from '../../../../../cloud-foundry/src/features/service-catalog/services.service';
 import { selectCfRequestInfo } from '../../../../../cloud-foundry/src/store/selectors/api.selectors';
+import { getIdFromRoute } from '../../../../../core/src/core/utils.service';
 import { entityCatalog } from '../../../../../store/src/entity-catalog/entity-catalog.service';
 import { CF_ENDPOINT_TYPE } from '../../../cf-types';
